refactor: reorganize repo into npm workspaces
