+++ packages/client/public/libs/three.core.js
