+++ packages/client/src/auth.ts
