+++ packages/server/tests/turret-traverse.test.ts
@@ -1,14 +1,14 @@
 // turret-traverse.test.ts
 // Summary: Verifies that only tank destroyers retain horizontal turret traverse limits.
 // Structure: spawn server -> fetch tank list -> assert non-TDs report 0 traverse and TDs keep their limit.
-// Usage: run with `npm test` which executes `node --test` against compiled output.
+// Usage: run with `npm test` which executes `node --test packages/server/dist/tests` after building the workspace.
 // ---------------------------------------------------------------------------
 import test from 'node:test';
 import assert from 'node:assert';
 import type { AddressInfo } from 'node:net';
 
 process.env.PORT = '0'; // use random available port
-const { server } = await import('../server/tanksfornothing-server.js');
+const { server } = await import('../src/tanksfornothing-server.js');
 await new Promise<void>((resolve, reject) => {
   const onError = (error: Error) => {
     server.off('error', onError);
