+++ packages/shared/src/index.ts
@@ -1,0 +1,7 @@
+// index.ts
+// Summary: Public entry point for the shared workspace, re-exporting reusable math and terrain helpers.
+// Structure: Named exports for terrain-noise utilities to support both client and server packages.
+// Usage: import { generateGentleHills } from '@tanksfornothing/shared';
+// ---------------------------------------------------------------------------
+
+export { generateGentleHills } from './terrain-noise.js';
