+++ tsconfig.json
@@ -1,8 +1,9 @@
 {
   "files": [],
   "references": [
-    { "path": "./tsconfig.server.json" },
-    { "path": "./tsconfig.client.json" },
+    { "path": "./packages/shared" },
+    { "path": "./packages/server" },
+    { "path": "./packages/client" },
     { "path": "./tsconfig.admin.json" }
   ]
 }
